requirements.txt
Django==6.0
Pillow==10.0.0
python-decouple==3.8
pytz==2025.1
sqlparse==1.8.3
